feat: finalizacion de separacion de roles y permisos

+++ src/main/java/fap/SistemaGestionEducativa/service/impl/EvaluacionServiceImpl.java
@@ -81,7 +81,7 @@ public class EvaluacionServiceImpl implements EvaluacionService {
 
         evaluacion.setNombre(request.getNombre());
         evaluacion.setTipo(request.getTipo());
-        evaluacion.setPeso(BigDecimal.valueOf(request.getPeso()));
+        evaluacion.setPeso(request.getPeso());
         evaluacion.setFecha(request.getFecha());
         evaluacion.setCurso(curso);
